Add ability to turn LSTM into a C struct.
I tried this with OpenCL and got it to work (passing the LSTM to OpenCL
and running the same code there).

Unfortunately, the GPU OpenCL was not much faster even on RTX 3090 Ti
than a fully saturated Ryzen 3950X.

Still, I may go back to that and might use this code again.